[Bug修复]: 修复弹幕cid过长导致的保存失败

+++ data_component/src/main/java/com/xyoye/data_component/data/DanmuData.kt
@@ -18,7 +18,7 @@ data class DanmuData(
 @Parcelize
 @JsonClass(generateAdapter = true)
 data class DanmuContentData(
-    val cid: Int,
+    val cid: Long,
     val p: String = "",
     val m: String = ""
 ) : Parcelable
